chore: Create Corpus (#2)
* Create Corpus

Cleaned and filtered the Corpus for further visualisation.

* Add proof of concept

Added a KML(Keyhole Markup Language) file in the directory POC which
will work as a proof of concept for the proposed soil assessment and
farm planner system.

* Create Readme

Created a Readme.rtf file which contains the metadata of every
attribute/column in the soil_data corpus, also changed the format of
the soil_data from xlsx to csv.

* Docs: Update README

added Resource section where different types of resourves used
throughout the project are cited.

* Change Readme format

changed the Readme file which consisted of the metadata of the
attributes of soil_data corpus from rtf to md.

* Update README

Added an embedded HTML link for the proof of concept of the GIS based
soil assessment farm planner .